+++ src/main/java/ua/rd/pizzaservice/services/OrderService.java
@@ -7,7 +7,7 @@ import ua.rd.pizzaservice.domain.Order;
  * @author Anton_Mishkurov
  */
 public interface OrderService {
-    Order placeNewOrder(Customer customer, Integer... pizzasID);
+    Order placeNewOrder(Customer customer, Long... pizzasID);
 
     void saveOrder(Order newOrder);
